staff_management/requirements.txt
asgiref==3.8.1
auditlog==0.0.5
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.2
django-filter==24.3
django-jsonfield==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular-sidecar==2024.11.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
filelock==3.16.1
httpie==3.2.3
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.6.0
pyttsx3==2.88
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.3
six==1.16.0
SpeechRecognition==3.10.1
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.27.0
